fix german long vowel preprocessing issue
In a word like Haar, the aar gets rewritten to aaa because the
r gets deleted, leaving aa, and then the
"open syllable lengthening" rewrite rules get applied to the a#
yielding aa; i.e. the aa# becomes aaa.
In general, the open syllable lengthening rule should not apply
to a vowel which is already lengthened.