fix vowel lengthening issue for german -- again
Again, the vowel lengthening rule should not apply to already long vowels.
What we had was ö can lengthen to öo via the rule ö -> öo
and then öo can become öoo via the rule o-> oo. Not good.
Also, öo needed to be added to the grapheme-phoneme map for german,
so that it gets converted into the IPA notation for long vowels.